requirements.txt
pywebview>=4.0
PyInstaller>=5.0
pystray>=0.19.0
Pillow>=9.0.0
pynput>=1.7.0